Fix/#26: 렌트 대기 예약 수정
컨테이너 삭제에 따른 코드 수정

반복되는 구조의 최적화
